split up md2conf.py into a start script and a class file
